TST: raise flag_crit in test_simulated
the test_simulated file is very short, approx 20 seconds. So a few epochs with blinks will be outliers and accont for more
than 20 percent of total epochs, causing the channel to be flagged. Raise the flag_crit to 30 percent to be more liberal so that blinks in a channel to cause it to be flagged